tests/wpt/web-platform-tests/content-security-policy/plugin-types/plugintypes-mismatched-url.html

<!DOCTYPE html>
<html>

<head>
  <meta http-equiv="Content-Security-Policy" content="plugin-types application/pdf;">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
  <script>
    var t = async_test('Should not load the object because its declared type does not match its actual type');
    window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
      assert_equals(e.violatedDirective, "plugin-types");
      assert_equals(e.blockedURI, "");
    }));
  </script>

  <object type="application/pdf" data="/content-security-policy/support/media/flash.swf"></object>
</body>

</html>
